Add correciones vistas transacciones

+++ Interfaces/IRepositorioTransacciones.cs
@@ -4,7 +4,7 @@ namespace ManejoPresupuestoNetCore.Interfaces
 {
     public interface IRepositorioTransacciones
     {
-        Task Actualizar(Transaccion transaccion, decimal montoAnterior, int cuentaAnterior);
+        Task Actualizar(Transaccion transaccion, double montoAnterior, int cuentaAnterior);
         Task Borrar(int id);
         Task Crear(Transaccion transaccion);
         Task<Transaccion> ObtenerPorId(int id, int usuarioId);
